[LPT] ConcatTransformation: removed axis limitation + added tests for covering only transformation logic
