+++ src/tests/functional/inference_engine/lp_transformations/concat_with_different_precision_on_children.cpp
@@ -127,6 +127,7 @@ public:
         result <<
             LayerTransformation::getTestCaseNameByParams(precision, inputShape, testValues.params) << "_" <<
             (testValues.multiChannels ? "multiChannels_" : "notMultiChannels_") <<
+            "axis" << testValues.axis <<
             testValues.actual << "_" <<
             testValues.result << "_";
         return result.str();
@@ -147,7 +148,7 @@ const std::vector<ngraph::element::Type> precisions = {
 const std::vector<ngraph::PartialShape> shapes = {
     { 1, 3, 10, 10 },
     { 4, 3, 10, 10 },
-    { Dimension::dynamic(), 3, Dimension::dynamic(), Dimension::dynamic() }
+    { -1, 3, 10, -1 }
 };
 
 const std::vector<ConcatTransformationTestValues> testValues = {
@@ -171,7 +172,7 @@ const std::vector<ConcatTransformationTestValues> testValues = {
             { ngraph::element::f32, {}, { 0.01f } }
         }
     },
-    // U8 with unsupported axis
+    // U8 concatenation by spatial dimension
     {
         LayerTransformation::createParamsU8I8(),
         false,
@@ -182,13 +183,13 @@ const std::vector<ConcatTransformationTestValues> testValues = {
         },
         {
             { 256ul, ngraph::Shape({}), {0.f}, {2.55f}, {0.f}, {255.f} },
-            { 256ul, ngraph::Shape({}), {0.f}, {2.55f / 2.f}, {0.f}, {255.f} },
+            { 256ul, ngraph::Shape({}), {0.f}, {2.55f / 2.f}, {0.f}, {128.f} },
             ngraph::element::u8,
-            {{ngraph::element::f32}, {}, {0.01f}},
-            {{ngraph::element::f32}, {}, {0.005f}},
+            {},
+            {},
             ngraph::element::f32,
-            {{}, {}, {}},
-            {{}, {}, {}}
+            { ngraph::element::f32, {}, { 0.01f } },
+            { ngraph::element::f32, {}, { 0.01f } }
         }
     },
     // I8
