+++ src/tests/ngraph_helpers/lpt_ngraph_functions/include/lpt_ngraph_functions/concat_function.hpp
@@ -17,6 +17,15 @@ namespace subgraph {
 
 class ConcatFunction {
 public:
+    static std::shared_ptr<ov::Model> get(
+        const ov::element::Type inputPrecision,
+        const ov::element::Type deqPrecision,
+        const std::vector<ov::PartialShape>& inputShapes,
+        const std::vector<DequantizationOperations>& dequantizationsBefore,
+        const std::int64_t concatAxis,
+        const ov::element::Type precisionAfter = ov::element::undefined,
+        const DequantizationOperations& dequantizationAfter = {});
+
     static std::shared_ptr<ngraph::Function> getOriginal(
         const ngraph::element::Type precision,
         const ngraph::PartialShape& inputShape,
@@ -106,6 +115,15 @@ public:
         const ngraph::Shape& reshapeOutputShape,
         const FakeQuantizeOnData& fqOnData1,
         const FakeQuantizeOnData& fqOnData2);
+
+    static std::shared_ptr<ov::Model> getReference(
+        const ov::element::Type dequantizationPrecision,
+        const ov::element::Type precisionBefore,
+        const std::vector<ov::PartialShape>& inputShapes,
+        const std::vector<DequantizationOperations>& dequantizationsBefore,
+        const ov::element::Type precisionAfter,
+        const DequantizationOperations& dequantizationAfter,
+        const std::int64_t concatAxis);
 
     static std::shared_ptr<ngraph::Function> getReference(
         const ngraph::element::Type precision,
